Has background and ball (static)
In this upload I was able to mess with the Index to upload images and clean up the main code. We have a baseball that starts at home plate of the background image and the background is a cool picture of a stadium with fireworks. Thought this is a great starting point to begin animating the ball.