Some updates on the ball movement
Hey Jake, thanks for doing the gravity fall. I utilized the addEventListener method because it is way easier than setting up a setInterval with onclick functions. Much more simplistic when we move forward. Any button on the keyboard should make the ball jump (space bar, arrows, the letter 6, etc). In my opinion, we should make it jump a little higher but for now lets just get the fielders in the mix so we can see if we need to alter jump/gravity variables. I like your idea of adding a sound to the clicks. I haven't done that yet but you can update if you want at anytime. Lastly, big issue we need to fix is the window; live-server leads to a window that you can scroll on. I will look into fixing this