Added upward movement to the ball
Hey Jake, thanks for setting up the gravity stuff. First thing, I used the addEventListener method to make the ball jump up when you click. In my research, it is much easier to use than setInterval method and doesn't look as bad in our code as well. I like your idea of adding sound when the ball jumps but did not add that in yet. Last thing is that the live server window is huge and you can scroll on it; we need to fix this. I will look into it