this has acceleration added
movement is a little gimmicky. feel free to not use this version or mess with the variables. forgot to comment them out but should be line 26 and 37 for new ballDY variable and also gravity now is really important because of new equation on line 88 (creates an acceleration without having to make a new variable). mess around with the gravity and moveUp function variable for ballDY and see if you get something you like